+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New Features
 
 - `init()` added to assist with setting up a project to use envsetup (#21)
+- `build_from_config()` added to add ability to also use config to create your directories (#25)
 - `rprofile()` will automatically store your configuration file in a standard location with a standard name, allowing `library()` to use this to re-assign autos
 
 ## Breaking Changes